Removed CMS-specific contents

+++ lib/dataformat/blockreplica.py
@@ -2,12 +2,15 @@ import time
 
 from exceptions import ObjectError
 from block import Block
+from _namespace import customize_blockreplica
 
 class BlockReplica(object):
     """Block placement at a site. Holds an attribute 'group' which can be None.
     BlockReplica size can be different from that of the Block."""
 
-    __slots__ = ['_block', '_site', 'group', 'is_complete', 'is_custodial', 'size', 'last_update', 'files']
+    __slots__ = ['_block', '_site', 'group', 'is_custodial', 'size', 'last_update', 'file_ids']
+
+    _use_file_ids = True
 
     @property
     def block(self):
@@ -17,38 +20,65 @@ class BlockReplica(object):
     def site(self):
         return self._site
 
-    def __init__(self, block, site, group, is_complete = False, is_custodial = False, size = -1, last_update = 0):
+    @property
+    def num_files(self):
+        if self.file_ids is None:
+            return self.block.num_files
+        else:
+            return len(self.file_ids)
+
+    def __init__(self, block, site, group, is_custodial = False, size = -1, last_update = 0, file_ids = None):
+        # Creater of the object is responsible for making sure size and file_ids are consistent
+        # file_ids should be a tuple of (long) integers or LFN strings, latter in case where the file is not yet registered with the inventory
+
         self._block = block
         self._site = site
         self.group = group
-        self.is_complete = is_complete
         self.is_custodial = is_custodial
-        if size < 0 and type(block) is not str:
-            self.size = block.size
-        else:
-            self.size = size
         self.last_update = last_update
 
-        # set of File objects for incomplete replicas (not implemented)
-        self.files = None
+        if size < 0 and type(block) is not str:
+            self.size = block.size
+            self.file_ids = None
+        elif size == 0:
+            self.size = 0
+            if BlockReplica._use_file_ids:
+                self.file_ids = tuple()
+            else:
+                self.file_ids = None
+        else:
+            self.size = size
+
+            if file_ids is None or not BlockReplica._use_file_ids:
+                self.file_ids = None
+            else:
+                # some iterable
+                tmplist = []
+                for fid in file_ids:
+                    if type(fid) is str:
+                        tmplist.append(self._block.find_file(fid, must_find = True).id)
+                    else:
+                        tmplist.append(fid)
+    
+                self.file_ids = tuple(tmplist)
 
     def __str__(self):
-        return 'BlockReplica %s:%s (group=%s, is_complete=%s, size=%d, last_update=%s)' % \
+        return 'BlockReplica %s:%s (group=%s, size=%d, last_update=%s)' % \
             (self._site_name(), self._block_full_name(),
-                self._group_name(), self.is_complete, self.size,
+                self._group_name(), self.size,
                 time.strftime('%Y-%m-%d %H:%M:%S UTC', time.gmtime(self.last_update)))
 
     def __repr__(self):
-        return 'BlockReplica(%s,%s,%s,%s,%s,%d,%d)' % \
+        return 'BlockReplica(%s,%s,%s,%s,%d,%d,%s)' % \
             (repr(self._block_full_name()), repr(self._site_name()), repr(self._group_name()), \
-            self.is_complete, self.is_custodial, self.size, self.last_update)
+            self.is_custodial, self.size, self.last_update, repr(self.file_ids))
 
     def __eq__(self, other):
         return self is other or \
             (self._block_full_name() == other._block_full_name() and self._site_name() == other._site_name() and \
             self._group_name() == other._group_name() and \
-            self.is_complete == other.is_complete and self.is_custodial == other.is_custodial and \
-            self.size == other.size and self.last_update == other.last_update)
+            self.is_custodial == other.is_custodial and self.size == other.size and \
+             self.last_update == other.last_update and self.file_ids == other.file_ids)
 
     def __ne__(self, other):
         return not self.__eq__(other)
@@ -59,11 +89,7 @@ class BlockReplica(object):
         if self._site_name() != other._site_name():
             raise ObjectError('Cannot copy a replica at %s into a replica at %s' % (other._site.name, self._site_name()))
 
-        self.group = other.group
-        self.is_complete = other.is_complete
-        self.is_custodial = other.is_custodial
-        self.size = other.size
-        self.last_update = other.last_update
+        self._copy_no_check(other)
 
     def embed_into(self, inventory, check = False):
         try:
@@ -86,7 +112,7 @@ class BlockReplica(object):
         replica = block.find_replica(site)
         updated = False
         if replica is None:
-            replica = BlockReplica(block, site, group, self.is_complete, self.is_custodial, self.size, self.last_update)
+            replica = BlockReplica(block, site, group, self.is_custodial, self.size, self.last_update, self.file_ids)
     
             dataset_replica = site.find_dataset_replica(dataset, must_find = True)
             dataset_replica.block_replicas.add(replica)
@@ -147,7 +173,7 @@ class BlockReplica(object):
                 site_partition.replicas.pop(dataset_replica)
 
         dataset_replica.block_replicas.remove(self)
-        if unlink_dataset_replica and len(dataset_replica.block_replicas) == 0:
+        if unlink_dataset_replica and not dataset_replica.growing and len(dataset_replica.block_replicas) == 0:
             dataset_replica.unlink()
 
         self._block.replicas.remove(self)
@@ -158,6 +184,54 @@ class BlockReplica(object):
     def delete_from(self, store):
         store.delete_blockreplica(self)
 
+    def is_complete(self):
+        return self.size == self.block.size
+
+    def files(self):
+        block_files = self.block.files
+        if self.file_ids is None:
+            return set(block_files)
+        else:
+            return set(f for f in block_files if f.id in self.file_ids)
+
+    def add_file(self, lfile):
+        # Note: cannot be used with a file that is just created - it doesn't have an ID until it's registered with the inventory store!
+
+        if lfile.block != self.block:
+            raise ObjectError('Cannot add file %s (block %s) to %s', lfile.lfn, lfile.block.full_name(), str(self))
+
+        if self.file_ids is None:
+            # This was a full replica. A new file was added to the block. The replica remains full.
+            return
+        else:
+            file_ids = set(self.file_ids)
+
+        file_ids.add(lfile.id)
+
+        if len(file_ids) == self.block.num_files:
+            self.file_ids = None
+        else:
+            self.file_ids = tuple(file_ids)
+
+        self.size += lfile.size
+
+    def delete_file(self, lfile):
+        # Note: cannot be used with a file that is just created - it doesn't have an ID until it's registered with the inventory store!
+
+        if lfile.block != self.block:
+            raise ObjectError('Cannot delete file %s (block %s) from %s', lfile.lfn, lfile.block.full_name(), str(self))
+
+        if self.file_ids is None:
+            file_ids = [f.id for f in self.block.files]
+        else:
+            file_ids = list(self.file_ids)
+
+        # Let remove() raise ValueError if the file id is not found
+        file_ids.remove(lfile.id)
+
+        self.file_ids = tuple(file_ids)
+        self.size -= lfile.size
+
     def _block_full_name(self):
         if type(self._block) is str:
             return self._block
@@ -166,19 +240,19 @@ class BlockReplica(object):
 
     def _block_real_name(self):
         if type(self._block) is str:
-            return self._block[self._block.find('#') + 1:]
+            return Block.to_real_name(Block.from_full_name(self._block)[1])
         else:
             return self._block.real_name()
 
     def _block_name(self):
         if type(self._block) is str:
-            return Block.to_internal_name(self._block_real_name())
+            return Block.from_full_name(self._block)[1]
         else:
             return self._block.name
 
     def _dataset_name(self):
         if type(self._block) is str:
-            return self._block[:self._block.find('#')]
+            return Block.from_full_name(self._block)[0]
         else:
             return self._block.dataset.name
 
@@ -193,3 +267,13 @@ class BlockReplica(object):
             return self.group
         else:
             return self.group.name
+
+    def _copy_no_check(self):
+        self.group = other.group
+        self.is_custodial = other.is_custodial
+        self.size = other.size
+        self.last_update = other.last_update
+        if other.file_ids is None:
+            self.file_ids = None
+        else:
+            self.file_ids = tuple(other.file_ids)
